+++ apps/shared/avifutil.h
@@ -73,7 +73,7 @@ avifAppFileFormat avifReadImage(const char * filename,
                                 struct y4mFrameIterator ** frameIter);
 
 // Removes a single trailing null character from the image->xmp, if there is exactly one.
-void avifFixImageXMP(avifImage * image);
+void avifImageFixXMP(avifImage * image);
 
 // Used by image decoders when the user doesn't explicitly choose a format with --yuv
 // This must match the cited fallback for "--yuv auto" in avifenc.c's syntax() function.